Create photon-targeting JNI
