+++ build.gradle
@@ -1,12 +1,15 @@
 import edu.wpi.first.toolchain.*
 
 plugins {
+    id "java"
+    id "cpp"
     id "com.diffplug.spotless" version "6.24.0"
     id "edu.wpi.first.NativeUtils" version "2024.6.1" apply false
     id "edu.wpi.first.wpilib.repositories.WPILibRepositoriesPlugin" version "2020.2"
     id "edu.wpi.first.GradleRIO" version "2024.3.2"
     id 'edu.wpi.first.WpilibTools' version '1.3.0'
     id 'com.google.protobuf' version '0.9.4' apply false
+    id 'edu.wpi.first.GradleJni' version '1.1.0'
 }
 
 allprojects {
