+++ photon-targeting/src/test/native/cpp/PacketTest.cpp
@@ -16,6 +16,7 @@
  */
 
 #include <chrono>
+#include <vector>
 
 #include <units/angle.h>
