+++ photonlib-python-examples/poseest/tests/pyfrc_test.py
@@ -2,5 +2,3 @@
     This test module imports tests that come with pyfrc, and can be used
     to test basic functionality of just about any robot.
 """
-
-from pyfrc.tests import *
